Separate matching logic into module
